move setting state into component
